Back Power Protection Diode
Some USB host controller (specifically, the one on the Raspberry Pi 4)
can detect if a device is being back powered through the D+ line (that
is, drawing power through D+ when the 5V supply is off), and if so will
shut down the port. Add a diode on the D+ line to prevent power from
being drawn through it.