Use buffer on joystick and aux keys
There was a problem where joystick or auxiliary key inputs would result
in keyboard keys being detected when they were not actually pressed.
While this is accurate to the actual Commodore 64 keyboard and joystick
input, having it happen on this USB device meant that it could not be
correctly disambiguated by the emulator. As such, the board is
redesigned to use a 74HCT541 tri state buffer chip to isolate the
Joystick and auxiliary inputs from the keyboard. It is still possible
that keyboard inputs can incorrectly show up as joystick movement, but
this is much less likely to cause a problem.

With the board redesign, the clearance around the keyboard header was
also increased to allow it to fit, and the USB connector was changed
from a micro-USB connector to a USB-B connector (I had too much trouble
hand soldering the micro USB connector).